+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Make workflow visual more compact to prevent overflow
 ---
 
 ## [0.2.0-alpha.9] - 2026-02-02